fix(seed): remove float64 cast from RateNight (domain expects int)

+++ services/backend-core/cmd/seed/main.go
@@ -99,7 +99,7 @@ func main() {
 				Type:       roomType,
 				BedType:    strPtr(bedType),
 				Status:     status,
-				RateNight:  float64(rate),
+				RateNight:  rate,
 			}); err != nil {
 				logger.Warn("room insert failed", slog.String("number", roomNum), slog.String("error", err.Error()))
 			} else {
